Fixed the UI and changed the application to app routing
