+++ app/api/hello.ts
